+++ src/app/ui/mod.rs
@@ -1,4 +1,5 @@
 pub(super) mod debug;
+pub(super) mod dsp_widgets;
 pub(super) mod editor;
 pub(super) mod effect_graph;
 pub(super) mod export_settings;